Move downsample to data_movement folder and align as per example

+++ tests/tt_eager/python_api_testing/unit_testing/misc/test_downsample.py
@@ -146,7 +146,7 @@ def test_run_downsample(
         ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1
     )
     # Run downsample op
-    A_downampled_sharded = ttnn.downsample(A_sharded, downsample_params, output_dtype=dtype)
+    A_downampled_sharded = ttnn.downsample(A_sharded, downsample_params, dtype=dtype)
     A_downsampled = ttl.tensor.sharded_to_interleaved(
         A_downampled_sharded,
         ttl.tensor.MemoryConfig(ttl.tensor.TensorMemoryLayout.INTERLEAVED, ttl.tensor.BufferType.L1),
